Rename `scb` folder to `pxWeb` because those functions can be used for other sites implementing the PxWeb API
