+++ src/lib/pxWeb/getTableDetails.ts
@@ -1,4 +1,4 @@
-import { PxWebApiV2TableDetails } from "./PxWebApiV2Types";
+import { PxWebApiV2TableDetails } from "./pxWebApiV2Types";
 
 export async function getTableDetails(tableId: string, language: string = 'sv') {
   const url = new URL(`https://api.scb.se/ov0104/v2beta/api/v2/tables/${tableId}/metadata`);
